pyproject.toml
[tool.poetry]
name = "drf-realworld"
version = "0.1.0"
description = "Example DRF codebase that adheres to the RealWorld API spec."
authors = ["Ramanathan Vishnampet <ramanathan.vishnampet@exxonmobil.com>"]

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
djangorestframework = "^3.12.2"
dj-database-url = "^0.5.0"
drf-yasg = "^1.20.0"
python-decouple = "^3.4"
whitenoise = "^5.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}
pylint = "^2.6.0"
mypy = "^0.800"
isort = "^5.7.0"
pydocstyle = "^5.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88